небольшой рефактор добавил просмотр истории изменения записей, убрал возможность переходить по страичкам не авторизованным, добавил небольшое разграничение по ролям, добавил поиск по дате от и до, добавил время послденего захода пользователей

+++ public_html/app/views/user/index.php
@@ -4,36 +4,34 @@
 $title = 'Пользователи';
 ob_start();
 ?>
-<h1>User list</h1>
-<a href="index.php?page=users&action=create" class="btn btn-success-">Create User</a>
+<h1>Список пользователей</h1>
+<a href="index.php?page=register" class="btn btn-info">Создать пользователя</a>
 <table class="table">
     <thead>
     <tr>
-        <th scope="col">#</th>
-        <th scope="col">Username</th>
-        <th scope="col">Email</th>
-        <th scope="col">Email verification</th>
-        <th scope="col">Is admin</th>
-        <th scope="col">Role</th>
-        <th scope="col">Is active</th>
-        <th scope="col">Last login</th>
+        <th scope="col">ID</th>
+        <th scope="col">Имя</th>
+        <th scope="col">Логин</th>
+        <th scope="col">Роль</th>
+        <th scope="col">Последний вход</th>
     </tr>
     </thead>
     <tbody>
     <?php foreach ($users as $user): ?>
         <tr>
-            <td><?php echo $user['id']; ?></td>
-            <td><?php echo $user['username']; ?></td>
-            <td><?php echo $user['email']; ?></td>
-            <td><?php echo $user['email_verification'] ? 'Yes': 'No'; ?></td>
-            <td><?php echo $user['is_admin'] ? 'Yes' : 'No'; ?></td>
-            <td><?php echo $user['role']; ?></td>
-            <td><?php echo $user['is_active'] ? 'Yes' : 'No'; ?></td>
-            <td><?php echo $user['last_login']; ?></td>
-            <td>
-                <a href="index.php?page=users&action=edit&id=<?php echo $user['id'];?>" class="btn btn-primary">Edit</a>
-                <a href="index.php?page=users&action=delete&id=<?php echo $user['id'];?>" class="btn btn-danger">Delete</a>
-            </td>
+            <td><?php echo htmlspecialchars($user['id']); ?></td>
+            <td><?php echo htmlspecialchars($user['username']); ?></td>
+            <td><?php echo htmlspecialchars($user['email']); ?></td>
+            <td><?php echo htmlspecialchars($user['role']); ?></td>
+            <td><?php echo htmlspecialchars($user['last_login']); ?></td>
+            <?php if (isset($_SESSION['user_id']) and isset($_SESSION['user_role']) and $_SESSION['user_role'] ==3){
+                ?>
+                <td>
+                    <a href="index.php?page=users&action=edit&id=<?php echo $user['id'];?>" class="btn btn-primary">Edit</a>
+                    <a href="index.php?page=users&action=delete&id=<?php echo $user['id'];?>" class="btn btn-danger">Delete</a>
+                </td>
+                <?php
+            } ?>
         </tr>
     <?php endforeach; ?>
     </tbody>
